Update triggers sql
This is a partial update of the triggers sql. I copied the generated triggers file
back from live and committed the 'obvious bits'. I really just wanted to do a noise-reducing
pass so I've only committed :

1) addition of obvious tables (search_display, translation)
2) removal of 'all that utf8 stuff' e.g IFNULL(NEW.,_utf8'')
- this was added because it allows case sensitive comparison but
got reverted because it caused a regression on some mysql versions. We
don't deeply care but we want our committed version so be consisent with
what gets generated.

Bug: T283561
Change-Id: I31fe1215a818ebf59af0a68c23653c65ec002ea6