Prune out 'NA' from tsmart
Removes the NA data that we added in in our export seemingly. I've spotted it
in nick_name (which we don't import) and 'supplemental_address_1' - which we do. There
is no US state with an abbreviation 'NA' so I think it's OK to assume
that we should just strip 'NA'

Bug: T284784
Change-Id: I45fed9c630c13c6a242318ae22c97d179df47d2e