[Refactor] Move the insert contact function internals
This leaves the functionality of wmf_civicrm_message_contact_insert intact but moves the
functionality to an apiv4 style class/ action (WMFContact::save). The signature is not,
at this stage, standard but that can happen (or not) as we iterate.

My thinking is we might be able to extend the normal contact save action over time but
I think we need to feel our way into all of that.

In the interim - the thing
we are doing is separating out this code form the monolith of 'wmf-civicrm'
so we can group related code together rather than sorting by pasta-diameter

Bug: T283104

Change-Id: I89d48628d5e94323dbbec2751e41daf77fd8ef0e