Add api to fill wmf_donor table for 2021 donations
This ports over the previously used drush script. It gets rid of
the hard-coding so it stays up to date a bit more 'effortlessly'.
For this run of the script I'm not expecting to have to schedule
it so I've left porting over the batching out of scope for now.

This should be runable

drush --user=1 @wmff cvapi WMFDataManagement.updateWMFDonor version=4 maximumReceiveDate=2021-07-10

Bug: T280595
Change-Id: I413a998c96c6ac1b65b81654f102e8a74ad4bb9c