Update where join fields in apiv4 call
Colemans has updated the preferred syntax from contact.first_name to
contact_id.first_name when getting from a related entity.
The syntax we use is supposed to be deprecated. In fact
it turns out to be broken - which
I have logged as a bug https://lab.civicrm.org/dev/core/-/issues/2771
but we can update in the meantime

Bug: T285321
Change-Id: I28cf7011e2fafae4ef055144f5e27a7a3722ce00