New fix for 5.41 fails on render template
A few commits ago we added this version check so we wouldn't get
errors when  is called on the wmfContact::save api
in the believe it was what was needed for the 5.41 update.

However, after more digging.... the issue is
the exception changed. We were catching API_Exception
when the property wasn't set (& the __call was throwing that).

It changed to throw a CRM_Core_Exception

So, if we catch both we can undo the last change &
it still works on 5.41

I logged upstream as
https://lab.civicrm.org/dev/core/-/issues/2768

as ideally the exception would NOT change

Bug: T285321
Change-Id: Ia4f7017d8a39d3d1241505fda243e12a75e6f7b2