Remove switcheroo button from activities
To replicate this is is necessary to ensure an assigneee group is set under
Administer->System Settings->Activity Assignee Settings

Configuring a group restricts the autocomplete on the assignee field
when creating a new activity (eg. from a contact record 'new meeting').
With this patch the button in the phab image should not be present

Bug: T287644
Change-Id: Ife2a301faae8d83f130d919335ac79247fecbde8