Update civitoken
We really just use the address token that formats the address here and the date tokens.

There is quite a bit of change because we haven't updated for a long time but
it's mostly just convention/boilerplate. People have submitted
a few more tokens but we
probably don't care much

https: //github.com/eileenmcnaughton/nz.co.fuzion.civitoken/commits/master

Bug: T290050

Change-Id: I432ca6467a60d1bc2077d39fbec4a05c8edc220e