Add sql update to remove end dates where they shouldn't be there
I tried this on staging & it took around 20 seconds. I incremented the
extension version since there is an update - but I *think* it would
have run it anyway.

I ran the update locally in the UI going to the status page. There is a
way to call via the api - I just always forget if it is update or upgrade or
whatever

Bug: T283798
Change-Id: I1ac662aa67e10c7325186ae01561a6122656e2a5