Fix import to limit soft credit to orgs
The import Melanie is working on is failing because there are 5 contacts
with the name of the organization she is trying to soft credit to.

However, 1 is 'the org' and the other 4 are employees. It seems pretty
clear by the fact we are checking organization_name that our code is set up to
handle the possibility of an organization (not an individual) being soft credited.
Adding contact_type to the criteria should prevent the ambiguity

Bug: T283104
Change-Id: I8784570594ef8b12a44f0d2a0532c56fd7663cff