Move omnimail tests to the end
We have a new test problem when I attempt to update civicrm core code
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/civicrm/+/716113

The tests pass. However when the silverpop connector is destructed it
attempts to log out. The log out attempt relies on the guzzle
middleware returning the previously set up mock response.

I think the fails must be caused by a 'new' outgoing http
request 'using up' the response. I *think* moving the
omnimail tests to the end might allow us to flush this out.

Hopefully if we merge this the rc update will either
a) fail differently or
b) pass

If b then once it's merged we can try moving this back up the running
order until it recurs to find out where it's happening. If a
hopefully it is more diagnosable.

If c (neither) I'll still be stuck :-)
Bug: T290620
Change-Id: Ic69b1542774633b66b10bab1d33c92047af4449c

Change-Id: Id10d24074215cd4c33c951900792a0245e100f38