Drop indexes on wmf_donor fields identified as not required to be searchable.
Note that you should do a fresh build before pulling down the patch
as opposed to pulling this down and then doing a build.

Or if you have already done a build - make sure the schema_version
in civicrm_extension is less than 4207 so the update runs
and drops the indexes.

Run using drush @wmff cvapi Extension.update

Bug: T288721
Change-Id: I066cf33c40bd0317cadd6a4e9f32f676593788ba