Fix contact name pairs to not show up as forgettable
There is a bug at the moment where the ui forgetme action
fails to render the page if the contact in question has
a firstname equal to NULL

Internally the failing api call is replicabale via drush (where
9 is the id of a contact with a NULL first_name).

drush @wmff cvapi Contact.getrefcount id=9
is not of type String

This is failing because in order to display the pseudo-join between
civicrm_contact and contact_name_pair_family we declare a CRM_Core_Reference
for the join. This mimics the way a 'real live fk' would be represented but
it is a pseudo-FK. However, we are seeing the getrefcount action is trying
to crawl this 'join'. So rather than using the Basic CRM_Core_Reference
class I've added a class that extends it, but always returns 0
when getrefcount is called. I've documented in
phpdoc comments how to see the pseudojoin still works in
searchkit

Bug: T244811
Change-Id: I269bb9fe80a69f61253056275202b8d7de4de052