Switch our recurring fail mail to use the new methods
The main functionality of civi-data-translate is in core now with the following changes

table name is civicrm_translation (old = civicrm_strings)
entity name is Translation (old = Strings)
the new table has a status_id column which can be 'active' or 'draft'
whereas the old had 'is_active' but we didn't use it much.

We are no longer using the apiWrappers as we are
1) only saving ourselves from one place - the qa screen & there we
are just saving directly
2) relying on the core code to do the work for 'get'
- this is called from our Message\Render->loadMessageTemplate()
function
In order to render in a given language it is necessary to pass
the 'locale' not the language into the tokenProcessor when we render it.
We are still using our own custom Render function in the
civi-data-translate extension and core support for an api
is still 'settling down' so we will likely revist that in a few months.

Bug: T288698

Change-Id: I9b03ba11a51fc77ce0514317cde8bc29c1b41d5c