X-Files acquired by Disney
The BitPay test was depending on a single donation getting processed
with the email address fox.mulder@pm.me. The file had a lot of rows
with that address, but all the ones with invoice currency EUR were
failing. When we set a fallback conversion rate for EUR, they would
all be imported and the test would fail because there was no longer
an 'id' property on the result.

Here we change the test file to leave only a single row with the
searched email address, and also add display names for the people
in that csv with names set to the 'mouse list' of contacts to clean
up. The display_name 'Anonymous' is already in the mouse list so that
should take care of most of the csv rows which don't have names.

Change-Id: Ifb5443bd54c5405673c96d91e59bc97597bf391c