More BitPay cleanup
Don't add the found contact ID to the list because it'll fail to
delete and complain. But DO add the other email address to the list
as that's used as a display_name in the absence of fname/lname.

Also add a lil more diagnostic info in case of cleanup failure.

Change-Id: I544c3867785c5aac7fc795a2e3771e489905a725