Function extraction in test class
This extracts the function such that we only have to change one
code place to switch to civicrm. The use of the curly braces in the
sql is good drupal coding practice because drupal can handle possible db-prefixing.

However, we don't db-prefix those tables and we are moving them to the other
db so removing them now makes sense

Bug: T290253

Change-Id: Ic8ea3956af1ff823e26ad97f1a8440b2fa4867dd