Boilerplate updates in test loading
This contains a bit of boilerplate regeneration but the important
part is the top level changes to bootstrap-phpunit.php

We have a require_once in there to include the 'cv' utility. This ALSO
loads the namespace - but it does it just for the one extension and that extension is
not 'ideal' in that if we want to add things that we want to be
available to tests in other extensions we would put those in wmf_civicrm.
In particular I want some test helpers to go in there. I'm still pondering
a better option - maybe using test listeners - but this gets me past
my current namespacing issue - maybe

Change-Id: I7b6d8df05d888e557d15c4347d802ef3e69b567c