Use contribution.amount rather than total_amount in the templates
On digging into formatting I think there is a need when building up contributions to
have a formatted & an unformatted valude. The latter is used for calculations. On
reflection I feel like it's OK to reserve the db name for the raw value and
offer up a shorter one to the users (total_amount is a bit confusing but I try
to use the DB names unless there is a reason not to & I think offering a formatted version
counts as that

Bug: T290253
Change-Id: Ieeaf629049837a2e8f47de5996899f9872447cd9