Stop filtering out endowment
These are now filtered at the template level. Note that these rows aren't
really going to be used at some point - the benefit of pre-querying them
and storing them in the db is doubtful from a performance point of
view and bad from the point of view of code re-usability. However, that
is out of scope of this PR

Bug: T290253

Change-Id: I6345b0ce93cdff6797555b5997c4a993b38e1846