Switch render to use the workflow message template
Note that this means that the translations will be active by default on dev setups. This seems better
to me now that they are deployed to live - however it has required quite a bit of change to the unit
tests. In the Smashpig class that change is because we are now using the translation file rather
than the message template & need to manage switching it for the test.

Bug: T290253
Change-Id: I8ac2e32ff2b29929dd1a0fa5d9a8882fdbadddff