Add link to Acoustic from mailing events
In trying to dig into bug https://phabricator.wikimedia.org/T293587
and update the docs too I found it a pain that we had the contact identifier
but couldn't use it to to find the acoustic record.

This adds a link to look up the contact in Acoustic based on the contact_identifier in
mailing_provider_data table to the MailingEvents tab

This sql will add a viewable row to see at
https://wmff.localhost:32353/civicrm/contact/mailings/view?reset=1&force=1&cid=204

INSERT INTO civicrm.civicrm_mailing_provider_data (contact_identifier, mailing_identifier, email,
event_type, recipient_action_datetime, contact_id, is_civicrm_updated)
VALUES ('490521770780', ' ', 'a@b.com', 'SENT', '2021-11-24 21:39:16', '204', 0);

The alignment is a bit odd - I was gonna ignore that.

Bug: T293587

Change-Id: I79a77a914161991e71a7f73b138e653547e3cb6d