Assume import is an individual when donor advised fund specified
Has donor advised fund is only valid for individuals. Having that AND organization name populated
caused the whole csv to fail as it assumed an org. Opted to assume individual in this scenario.

Note the test data looked too realistic so I wound up checking it wasn't actual donor data
& then updating so that no-one would worry in future

Bug: T296861
Change-Id: Iddb7146182eec3df3a01a2a7986bbbf900810531