Run TY initial query via Civi query function
Using the Civi function for this query instead of the drupal function
means it can be intercepted by RPOW and run on the replica database.

It's a problem if this setup query is run on the primary while the
follow-on individual mailing details queries are run on the replica,
because the results of this query could include contribution IDs that
aren't yet replicated.

Bug: T296813
Change-Id: Iba7439c0fee4280b0971481b370ba3a241ae792e