Alter query to ensure email is unique
This allows us to call the MakeJob action more than once & only one
job will exist for each email.

I tested on staging and the final query for 2021 takes
3 mins 20 sec on an empty table
2 mins 50 sec when re-run - ie the table is already populated.

I altered the query to use contribution_recur as the base table for
efficiency but it was perhaps 20% faster.

Last year the MakeJob step was slow & stalled the queue processing but
we are only calculating the email list this year so it's not clear we will
need to stop queues (a 3 min query is not 'that' bad).

Also note that I added the join to the existing table so the query can be run multiple
times and it will add each relevant email a max of once. This felt safer given the
removal of job_id. I don't specifically see a need to run more than once but it
might be interesting to run at the end & see if new emails pop up (most
likely due to changes to the primary email). I do see a few missed from
last year but I didn't dig too much

Bug: T297159
Change-Id: I452c1ffb75fdfb9e149241b25534850581534128