Switch up flow - render, email & save per row
Now we have refactored we can re-organise this into a better flow. We aren't gaining
anything by rendering all the emails at once & it means that if it does fail during rendering
the rendered emails are not sent. Better to send as we go.

Also we can start passing contact IDs around again

Change-Id: I154806f03afe06a0cd601b6adb459c61d9daeeb1