Update naming convention on template files
The naming convention will eventually see the
thank you templates moved to
sites/default/civicrm/extensions/wmf-civicrm/msg_templates/thank_you

The convention in that folder structure is to have all files for
one workflow template in one folder (rather than html & subject).
It also uses a .txt suffix (I wondered about changing that convention,
which is also used in CiviCRM but it didn't seem to gain us much
and then we would wind up without a .txt suffix on .subject files
which is annoying).

Changing that required a change to our Templating loader - so
I made the change to all files loaded by that loader. I note that
the other way we will deal with templates in the other modules
is still under consideration - but not actively being addressed.

Is the exchange script even used now?

Bug: T299455

Change-Id: I568461b621106cc148acc6bc2237c2a573bdca02