Remove 'update' = never for reserved templates
At this stage users can't edit the default reserved templates &
they are not auto-synced with the code version = confusing...

Removing the 'update'=never key means that when we make local
updates the 'master' but not any
of the editable ones is updated. The master is
never emailed - but it can be seen in the UI
on the 'original' tab

I have updated the default version to what is currently on
production. This involves some minor changes to tests

Change-Id: If17e538d9314e1e629b32ea508c3eae75f4df4a4