Update extendedreports
The changes in this update are all very superficial - but there is
one that is important. The grant report has been fixed to
only install if CiviGrant is enabled. In 5.46 CiviGrant is moved to
an extension so installing it when the extension is not
installed (as we do) now causes a constraints error.

Other changes in this update are null changes to us

Bug: T294487
Change-Id: I7edb1e3316edc65ad2239ab480880d2adf1ba5fd