Update ContributionRecur.getoptions call
There is a change in 5.47 whereby the results
of getoptions return the label rather than the
name. This is 'more correct' but doesn't tie with
our usage

The patch updates our code to use apiv4 which is
more specific about what is done

The best way to test this is through the test RecurringQueueTest.testCreateDistinctContributions.

The value of processors should be the same as before - as long as
we don't have the updated CiviCRM - at which point it
would otherwise break
https://lab.civicrm.org/dev/core/-/issues/3075

Bug: T294487

Change-Id: I93d624234e8d2c9b6ffdbe95f3389b798b263dbe