Update original endowment letter
This updates the 'original' - but not the 'current' on prod - which still
needs to be UI-edited. If a site is rebuild with this (civibuild wmff create)
the site will get this as current too

Note that with this
patch https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/civicrm/+/770660
the updated version can be viewed as an example.

Bug: T302448
Change-Id: Ia0747922ddcba9bb5a7ceb9cc1859189ce888e85