Update preferred_language on new contribution
Back when we used to create a new contact record for each new
contribution, the merge process would pick up the language from
the most recent contribution.

In change I7ea9e112f2b020224f8fb7025ca0af52056e9420, we switched
to attaching the contribution to existing contacts that match the
name and email, and to update communication preferences according
to what come in with the new donation, but we didn't write code
to update the preferred language on the existing contact.

We limit this to just individual contacts, not organizations.
The getPreferredLanguage method is refactored to take the full $msg
array so it's easier to call from two places.

Bug: T308043
Change-Id: I04ca644cd25bf403102e3c0376735c2d140b00b1