Remove problematic import tests
The signature on various import functions has changed in the latest CiviCRM
code - causing the fails we see in
https://integration.wikimedia.org/ci/job/wikimedia-fundraising-civicrm-docker/8089/console

The authorization error is hard to fix - as we have to simulate a logged in user. In
addition the signature of the import function is different in the upgraded
version to the existing so it's really hard with our currenct separate-repo structure
to deal with this - since the test has to work on both civi versions.

This led me to question the value of the test and the legitimacy. The test is calling
core import classes that are not supported to be called from outside of core (yes
we do call them in targetsmart too and will need to update that code to use whatever
is agreed as a supported interface if we ever use it again) and the extension is
not interacting directly with the import code - rather the import is testing the
same code patch (the Contact.pre hook) that all the other bits of code are testing.

Hence I concluded it is actually OK to remove the test
Change-Id: I50fc000f717e2d96a33e7c2dd75e46e3b6e5b3ea

Change-Id: I785bc27ddbd93e10895b76f28d918def5b8af85d