Do not exclude CiviCRM files from git
This will specifically affect vendor - which we have checked in. We can
think about changing how we do vendor in the future but at this
stage we need to be ensuring updates are not skipped. Also
let's take a moment to enjoy the ease of a single repo
before starting on a project to make it more
correct but harder to work with

Change-Id: Ib1efcc110a878e5ac6cf83c32f5f8fe2004b5906