Reactivate cancelled recurrings on import
Adds a helper function to do the reactivation, calls it from the
donation queue consumer and the recurring queue consumer. This
is for the general case.

Note that there is a separate function for reactivating legacy PayPal
donations that were mistakenly cancelled (rewritten in parent patch)
but that function also sets an ID we don't want to set here.

Bug: T242158
Change-Id: I3d7949c80a323785fd58a928d15f411861552f64