Add SQL for contribution_source
This table exists on production but isn't checked in to any of our
current repositories. It was created by the old ContributionTracking
mediawiki extension that used to manage the contribution_tracking
table in the drupal database. The table SQL checked in here is taken
from the current state of production.

In the upgrade script we only create the table, not add the triggers.
It was a pain to try to get that trigger file sourced because it
messes with delimiters, and I want to drop the triggers with the next
patch anyway.

Bug: T316898
Change-Id: I548c15b5e94c9da170843e4284a0b90499b9c3a2