Move contribution_source population into code
DEPLOY NOTE: need to stop the contribution tracking QC job and
drop triggers on the contribution_tracking table before pushing this
out.

Now that all the changes to this table go through the c_t queue
consumer, we can parse the contribution_source columns out in code
and not in this mysterious trigger.

Bug: T316949
Change-Id: I670770b069580f93fd67dbc42854364df5035f61