Add Adyen data deletion requests to forgetme
Adds ForgetMe and ShowMe actions on the Contribution entity which
send data deletion requests to any payment processors which support
the IDeleteDataProvider interface (so far just Adyen) and show a
brief summary of the contribution on the data export page.

To test, make some Adyen contributions, consume queues, and on the
donor record in Civi select Actions -> Forgetme.

Bug: T311741
Change-Id: Iefb21a069a43288b54efa4d32cbf3818ee0d3bc1