Adapt to exception changes in civi-upgrade
CiviCRM is phasing out the two api exception classes in favour of CRM_Core_Exception (as
they are not meaningfully different). It does this by using class aliasing in
the next release - which means we catch them more consistently but this throws out
our test expectations as the test was resigned to us not catching them & converting
to our own exceptions in erm exceptional cases.

This fixes our test the new expection and extends our catches temporarily - we
can start to only catch CRM_Core_Extension from next civi-bump

Bug: T316713

Change-Id: I6a726a92ab5bb95cc7104400f0b9959e6da2c212