Civi upgrade latest tarball
This deploys the latest tarball as of today. The code merged yesterday
had a couple of patches that were not merged into the tarball & were patched on
so this is the raw tarball

I also removed one that I pulled in
https://github.com/civicrm/civicrm-core/pull/24325
- I thought we would want to make that change so I could pull in the bits
that build on it - but I've changed my mind about wanting it in
core too & have a new plan....

This is exactly the same as the tarball less one thing -
In CRM_Core_Smarty it turns on e-notice for smarty in debug mode.
I don't normally comment that out but over big English we have a
four month period where we don't update - so we keep seeing
the notices but it's not worth fixing them
as they might already be fixed upstream. So it's an annoyance.

Maybe not more of an annoyance than coming so achingly close to
finally pushing out a 'stock' tarball & being two characters away from
doing it.....

Bug: T316713
Change-Id: Ie62cd820202e3147785eedc95ff001d9e917ce34