Fix for deduper angular error
When selecting a property from the 'Add clause' dropdown there is a
JS error in this file's getUrl function: deferreds is null.

Core expects all of the templates to come in with the
$http.get(moduleUrl) call. Looks like previous behavior was to do
a fallback lookup for a single template when it doesn't exist in
the template cache. The new behavior throws an error.

This template doesn't exist on the filesystem, so it isn't loaded
in the moduleUrl get. Stop referring to the non-existant template.

Bug: T318833
Change-Id: Idf9fba9ea21f5e2f69006ee1761aa4f6d25e2c90