Fix CancelInactives to set status
There is a difference between . & : in api4. The first is a join whereas the other
looks up the pseudoconstant. So, we were not actually updating the status when
cancelling contributions. This fixes. I also fixed some name-spacing issues in
other tests (that I tried to use for copy & paste purposes :-)

 Bug: T315102

Change-Id: I2b5683d43d161f32ad7911cd792a29251e12899a