Don't resolve multiple txns for one donor
Checks latest donation date and does not capture if within the day.
Note that this check will only pick up donations that have already
been consumed into Civi and would miss ones that have just been queued.

We could also cache (in Civi::statics) a list of email/name pairs
with donations captured in the current job run and check against that.

Note that this patch also does not check the most recent donation date
for messages with risk scores below the threshold. Those will be the
minority of messages, and the code change will be larger, so I will
do that in a follow-on patch.

Bug: T320578
Change-Id: Iddaacca214df20998d258e06f3d974ce79c42db0