Skip capture for past-day donors even with low risk
We are already not capturing pending transactions for matched
donors with a donation in the past day when the pending risk score
is above the review threshold. This patch extends that precautionary
measure to pending transactions with a risk score below the review
threshold. This could help when donations weren't captured on the
front end for some reason other than fraud filters.

Also filters the match to not include donors with is_deleted=1.

Bug: T320578
Change-Id: Iea7286231ae63a32a9742e30f75703544adc3b4d