Fix trigger handling of null or negative contributions
Contributions with an amount less than or equal to zero should not
could towards the first contribution date - but they were not
being excluded due to the typo fixed here.

The trigger change is not 'final' - we will regenerate triggers from
live - but I thought it might make it clearer including it.

Bug: T322189

Change-Id: I2982e2343a5944893d578b969a09a59ccfc3fd63