Add Organization to examples for message templates
This adds an extra example into the available message templates for
the Thank You messages - Thank You (Organization).

It is possible to see this works by adding the token
{contact.email_greeting_display} into the text.

Note however, that there is a leakage issue where it is
using the pseudo-contact-id to pick up 'real' values from the database.

I'm aware of that but digging into it separately & think this patch can
still be merged as is - as this part of the code will not be different
as a result

Bug: T306556
Change-Id: I039dad319305f8267f48c1f48deb58ba7b9c5218