Add support & test for Organization specific greeting in emails
This allows us to update the greeting block on live (which
I will do with Thea) to the following

{if $first_name and $last_name}
Dear {$first_name},
{elseif $contact_type === 'Organization' && $email_greeting_display}
{$email_greeting_display},
{else}
Dear donor,
{/if}

It has a unit test for the actual usage & should also preview in the UI.

Note that prod (but not our dev environments) has an email greeting template
for organizations and major gifts have ensured that email_greeting_display
is populated for organizations.

Ideally I would fix our dev environments to have the email greeting template like
prod but I hit a bug in the api https://github.com/civicrm/civicrm-core/pull/24938
(I fixed the bug upstream but lost the will to live when I considered backporting it
and THEN fixing our dev environments).

To manually configure look in Administer->Communications->Email Greeting format.
However, note that it is not a requirement to set that up to test this - you
can edit it through the UI for any individual contact
Bug: T306556
Change-Id: If0bcfd9c6b0148551dca09698286ac86b87ec153

Change-Id: Ie82624527e2c6d6903479344a25a65b8c253e3ec