Add fastArrayDecorator to mixinBoot & mixinScan caches
In investigating places where the Redis cache has been hit the attempts
to get stuff from the 'mixinBoot' and 'mixinScan' caches stood out as
being very high (not quite the highest - still digging there).

These are new caches (hence could account for a recent-ish slow-down)
and have not been given the 'fastArrayDecorator' - the `fastArrayDecorator`
is the class that stores values in a php-static cache for the
duration of the request. In theory it's a bit naughty
as it doesn't follow ttl and won't respond if a different
request has meddled with the cache. In practice
it performs waaaay better & the consistency issue is more theoretical
than practical - that is especially the case here where the caches reflect the files
on disk. It is a bit annoying though that they are 'yet new' caches - as
they will need to be cleared in a bunch of places ....

https://lab.civicrm.org/dev/core/-/issues/3813#note_84650

Bug: T322172

Change-Id: I3fe17744afe817cad5e251631a40b59139b7319d