Remove civicrm_initialize from thank_you_for_contribution
I notice that the thank_you job appears not to be using the array cache  - my
theory is that is because we keep calling the civicrm_initialize function.

civicrm_initialize is quite heavy, so could have other impacts. It should be called once
per run (erm to initialize). In the case of this function it is called from 5 places

- twice from tests (it should run fairly centrally in tests now & CI will tell us if not)
- once from our v3 thank_you_send api - this is only accessible once civi is initialized
so doesn't need this call.
- once from thank_you_batch_process - which calls civicrm_initialize() at the start of the function
- once from thank_you_batch_test - same as above

Bug: T322172

Change-Id: I6c1a304ca526b7cddf885920254f7e14064895d3