Fix Adyen GDPR forgetme calls for merged contacts
The API3 contribution_get call seems to have a bug when called with
params that only filter on contact_id, ['IN' => [1, 2]] and when one
of the two contacts has no contributions. Here we add a trivial
condition on the contribution table to filter out a spurious row.

To test, merge two contacts in the test database and add some
contributions to one of them (could be Adyen contribs via the
front end if you want to test the whole functionality, but you
can test that it fixes the bug described on the ticket just by
adding contribs in the Civi UI). Then go to the contact that remains
undeleted after the merge, click actions->forget me, scroll to the
bottom, and click the forget button.

Bug: T322796
Change-Id: I41af1d21963ff7fdd2a7972d307a2c946d4cb284