Stop changing recur currency on contribution edit
Recent versions of Civi try to keep the contribution_recur amount
and currency in sync with the amount and currency of an assocated
contribution row. Specifically, when we edit a single contribution
in USD that is associated with a contribution_recur that has a
different currency, the core hook
CRM_Contribute_BAO_Contribution::self_hook_civicrm_post will call
CRM_Contribute_BAO_ContributionRecur::updateOnTemplateUpdated and
edit the currency of the recur row to USD (and the amount to match).

We can prevent this by hooking into the pre-save event of the
ContributionRecur entity and unsetting all the values that would
be set. The tricky part is targeting our hook to only nullify these
particular edits, without needing to do an expensive stack
examination on each operation.

Long-term, we should probably switch to storing non-USD contributions
the way that Civi core expects us to, with the original contribution
right on the civicrm_contribution table and not off in our custom
table. That will need some big updates from analytics at the very
least, as well as plenty of updates in our own custom code

Bug: T320286
Change-Id: Ic55894963853dd7703f8776f581b6ec993846b56