Add new 'donations_total_messages_import_time' metric
This is to capture the actual time spent importing donations whilst
excluding the time spent detecting & rejecting duplicates, which
is currently a part of the existing $batchProcessingTimee metric.

The code for this is dropped in alongisde our other new dequeued
metric. I don't like it's current form but we're pressed for time.

Note: this needs refactoring when we have more time e.g.
we should move the indivdual stats into the ImportStatsCollector
and the overall stat into the custom DonationStatsCollector
exporting code. we should probably get rid of the existing
$batchProcessingTime var and use this new metric instead.

I also shouldn't be pushing up code at 11:23pm the night before
turn on the firehose either but hey ho!

Bug: T323273
Change-Id: I047ed308be23dbfae2ac9c6fbeb7cf92f4f053ae