Add on_hold information to the deduper interface
To test this it's easiest to first create 2 contacts with the same email
but it should be on hold for one of them. Then go to
civicrm/a/#/dupefinder/Contact and choose find duplicates (if they
have the same email then the default rule will work &
you don't have to select anything else).

The email of the on_hold one should be visibly on hold

Bug: T321843
Change-Id: I61320a188c7d50f52bf13b76e064047bea1953c7