Hack Partner back into contactBasic form
We had been using the Afform override methodolody to add
our Partner field to the afform that ships with deduper. The rest of the
form is generic / suitable to others so the theory was site-specific
fields should be added through override. However, in order to
get the new change for email on hold I had to
revert to stock and then ... I found I could no longer
add an override using the UI as I did last time.

I've created https://lab.civicrm.org/dev/core/-/issues/4040 to
discuss but in the meantime this adds the partner field to the deduper screen

Bug: T321843
Change-Id: Ib3df6594fcb0a4364065b8fcd7e0d7e0b0189685