Remove code to update offline receipts
This code is out of date enough I think it makes more sense to remove & log a phab
for a better future approach. This template is moved into the civicrm database in
the civicrm_msg_template table and is used when a user chooses 'send receipt'
on the edit contribution page. However, they have been told not to do this
after it failed to keep pace with our ThankYou letter. A better fix,
for which I raised a phab - https://phabricator.wikimedia.org/T325956, 
would be to send our thank you letter when they
check the box, or to remove the checkbox. Both of these are out
of scope for this change though

Bug: T270680
Change-Id: I998f9dafc0fef1063403169576ce805b4d80dddb