Move enabling CiviCampaign to extension
Note we set settings in 3 different places
- extensions/wmf-civicrm/Managed/Settings.php
- sites/default/wmf_settings_developer.json
- sites/default/wmf_settings.json

The first & last of these are settings that are present on
production and dev. The middle is dev only settings.

In general the approach we have been moving to for settings
that are also on prod is the first, which allows in-code comments (yay) and
works slightly differently - setting the default value rather than
actively setting the setting. However, for logging & enable_components
DB changes are made when the setting is toggled (which setting
the default won't do). I've added some commenting
and moved a couple of settings that do not actively 'do something'
and enable_components.

Bug: T270680
Change-Id: I3a5839bc2ff7dc0ea9c2693d926e45bb8c7ee322