Move declaration of Dedupe fishing net rule to managed entities
The original install code does two things
1) legacy cleanup relevant to production
2) creates a new dedupe rule called 'fishing net'

We can drop the first & this moves the second to be a managed entity.
It will create or rows in civicrm_dedupe_rule_group (1) and
civicrm_dedupe_rule (6) along with tracking rows in civicrm_managed.

On existing sites the dedupe_rule_group & dedupe_rule entries should
exist and only the tracking rows should be created when doing a reconcile
- the command for which is

drush @wmff cvapi Managed.reconcile version=4

I had to do a minor upstrem patch for this which is incorporated here

https://github.com/civicrm/civicrm-core/pull/25227

Bug: T270680
Change-Id: I2cd82f064db95b7a54aaf58d16937bf03d3504d1