Switch wmf_campaigns to use Factory method
To test
1) at this url https://wmff.localhost:32353/admin/config/wmf_campaigns/edit/Event
add an email address & enable or, for bonus points add a comma separated list of emails
with spaces to trim
2) check that the event notification shows 'yes' at https://wmff.localhost:32353/admin/config/wmf_campaigns/list
3) Create a contribution, in the 'Gift Data' accordian  set the 'Direct Mail Appeal' to 'Event' & save
4) check mail catcher - http://wmff.localhost:1080/

This is the same method called internally. It looks like the reason this
didn't get done last time is because we have this magic going on
where we get the mailer to handle multiple 'to' addresses which
the smtpmailer we use for almost everything else doesn't  do.

After some time grooming the other rabbits
in that nice warm underground place it occurred to me that
our usage of this functionality is so limited we
could just simplify to sending more than one email.....

I also figured I could standardise the getNotificationAddressFromKey
to handle the comma separated string - rather than deep in our magic class...
The only other place that calls that actually resolves it to Yes/No so
I tried to make that a little clearer too.

Bug: T326271
Change-Id: I76b19c6f7e3bd879b3906bb6912d57dcd4a5e724